+++ src/tag/parser/compile/scenario.cpp
@@ -76,6 +76,10 @@ namespace Invader::Parser {
     }
 
     void Scenario::pre_compile(BuildWorkload &workload, std::size_t tag_index, std::size_t struct_index, std::size_t) {
+        if(this->child_scenarios.size() == 0) {
+
+        }
+
         if(workload.disable_recursion) {
             return; // if recursion is disabled, doing any of this will be a meme
         }
@@ -226,14 +230,14 @@ namespace Invader::Parser {
 
         // If we have scripts, do stuff
         if(this->scripts.size() > 0) {
-            if(this->scripts.size() > 0) {
+            if(this->source_files.size() > 0) {
                 workload.report_error(BuildWorkload::ErrorType::ERROR_TYPE_WARNING, "Scenario tag has script data but no source file data", tag_index);
                 eprintf_warn("This is DEPRECATED and will not be allowed in some future version of Invader.");
                 eprintf_warn("To fix this, recompile the scripts");
             }
             else {
                 // TODO: Recompile scripts
-                workload.report_error(BuildWorkload::ErrorType::ERROR_TYPE_WARNING, "TODO: Implement script re-compiling", tag_index);
+                eprintf_warn("TODO: Implement script re-compiling");
             }
         }